# internal/api/modelapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lineage Model API",
    "version": "v1",
    "description": "Machine-facing Model API for the Lineage AI model registry: publish, resolve, fetch, and manage models, versions, artifacts, lineage, and deployments. Auth is handled by infrastructure; a trusted X-Lineage-Actor header is recorded for audit."
  },
  "servers": [
    {
      "url": "/",
      "description": "Model API port (:8081)"
    }
  ],
...
  "paths": {
...
    "/v1/models/{model}/versions/{version}/validations": {
      "parameters": [
        {
          "$ref": "#/components/parameters/model"
        },
        {
          "$ref": "#/components/parameters/version"
        }
      ],
      "post": {
        "operationId": "recordValidation",
        "tags": [
          "compliance"
        ],
        "summary": "Record a validation of a version",
        "description": "Appends. Audited as `validation.record`. A self-validation (validator = version author) is recorded with independenceEvidenced false, never refused.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidationInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Recorded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Validation"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          },
          "422": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "get": {
        "operationId": "listValidations",
        "tags": [
          "compliance"
        ],
        "summary": "A version's validation history and state",
        "description": "Newest first; superseded validations are returned too — what was believed before is the question an examiner asks.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationList"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/v1/models/{model}/versions/{version}/validations/{id}:clearConditions": {
      "parameters": [
        {
          "$ref": "#/components/parameters/model"
        },
        {
          "$ref": "#/components/parameters/version"
        },
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "The validation id."
        }
      ],
      "post": {
        "operationId": "clearValidationConditions",
        "tags": [
          "compliance"
        ],
        "summary": "Record that a conditional validation's conditions are met",
        "description": "Sets conditionsClearedAt, once. 409 `reason: not_conditional` on any other outcome, 409 `reason: already_cleared` on a second call. Audited as `validation.conditions_cleared`. No request body.",
        "responses": {
          "200": {
            "description": "Cleared",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Validation"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "model": {
        "name": "model",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "version": {
        "name": "version",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
    },
    "responses": {
      "Problem": {
        "description": "Error (RFC 9457)",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    },
    "schemas": {
...
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "code": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "details": {
            "type": "object"
          }
        }
      },
      "FactSource": {
        "type": "string",
        "enum": [
          "declared",
          "derived",
          "measured"
        ],
        "description": "How a value was obtained. declared = asserted by the publisher (attributed, not verified); derived = computed by a producer from the model; measured = reported by a runtime or harness, valid within its stated basis."
      },
...
      "StaleReason": {
        "type": "string",
        "enum": [
          "review_due_passed",
          "version_published_since",
          "production_changed_since",
          "derivation_since",
          "validation_expired",
          "unmonitored_in_production",
          "conditions_outstanding"
        ],
        "description": "Why a classification went out of date. `production_changed_since` reads updated_at, so editing the production version's description trips it — a deliberate false positive, since for a legal field erring toward 'take another look' is the right direction. For `mrm` rows: `validation_expired`, `version_published_since` (a version published after the validation), `unmonitored_in_production` (in production with no evaluation run since it got there), `conditions_outstanding`."
      },
...
      "MRMState": {
        "type": "string",
        "enum": [
          "untiered",
          "unvalidated",
          "stale",
          "current"
        ],
        "description": "The §20.7 ladder, computed on every read for a version's latest validation. `unvalidated` covers no validation and a latest outcome of `rejected` or `undetermined`. Neither `untiered` nor `unvalidated` is a kind of `stale`."
      },
      "ValidationOutcome": {
        "type": "string",
        "enum": [
          "approved",
          "conditional",
          "rejected",
          "undetermined"
        ],
        "description": "`conditional` is fit subject to `conditions`, which are then required. `undetermined` is reviewed but cannot conclude yet."
      },
      "ValidationInput": {
        "type": "object",
        "required": [
          "outcome"
        ],
        "description": "There is deliberately no validatedBy or validatedAt field. Both are server-set — the actor header and the server clock — and a request that supplies either is rejected with 400: the value of both is that the registry witnessed them.",
        "properties": {
          "outcome": {
            "$ref": "#/components/schemas/ValidationOutcome"
          },
          "scope": {
            "type": "string",
            "description": "What was validated, e.g. 'credit decisioning use only'."
          },
          "findings": {
            "type": "string",
            "description": "What the validator concluded."
          },
          "conditions": {
            "type": "string",
            "description": "Required non-empty when outcome is `conditional` (422, `details.field: conditions`)."
          },
          "validUntil": {
            "type": "integer",
            "format": "int64",
            "nullable": true,
            "description": "Epoch millis, must be in the future (400). Null means no expiry, which is surfaced rather than hidden."
          },
          "evidenceArtifactId": {
            "type": "string",
            "description": "The report itself: an artifact of kind DOC or METRICS on this version (400 otherwise)."
          }
        }
      },
      "Validation": {
        "type": "object",
        "required": [
          "id",
          "versionId",
          "outcome",
          "validatedAt",
          "independenceEvidenced",
          "source"
        ],
        "description": "One independent judgement on one version. Append-only: a re-validation is a new row, and the current answer is the latest.",
        "properties": {
          "id": {
            "type": "string"
          },
          "versionId": {
            "type": "string"
          },
          "outcome": {
            "$ref": "#/components/schemas/ValidationOutcome"
          },
          "scope": {
            "type": "string"
          },
          "findings": {
            "type": "string"
          },
          "conditions": {
            "type": "string"
          },
          "conditionsClearedAt": {
            "type": "integer",
            "format": "int64",
            "description": "Set once by :clearConditions. While a conditional validation's conditions are uncleared, the state is `stale` with `conditions_outstanding`."
          },
          "validUntil": {
            "type": "integer",
            "format": "int64"
          },
          "evidenceArtifactId": {
            "type": "string"
          },
          "validatedBy": {
            "type": "string",
            "description": "Server-set from the actor header."
          },
          "validatedAt": {
            "type": "integer",
            "format": "int64",
            "description": "Server-set."
          },
          "independenceEvidenced": {
            "type": "boolean",
            "description": "validatedBy is set and differs from the version's author. A flag, never a refusal: a one-person team or a shared service account trips it legitimately, and a human decides whether it is a finding."
          },
          "source": {
            "allOf": [
              {
                "$ref": "#/components/schemas/FactSource"
              }
            ],
            "description": "Always `declared`."
          }
        }
      },
      "ValidationList": {
        "type": "object",
        "required": [
          "items",
          "state"
        ],
        "description": "A version's validations, newest first, with the §20.7 state of this version against the model's `mrm` row. Not paginated.",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Validation"
            }
          },
          "state": {
            "$ref": "#/components/schemas/MRMState"
          },
          "staleReasons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StaleReason"
            }
          }
        }
      }
    }
  }
}